refactor: вынести логику Storage в абстрактный класс
